feat(prototype)：合约的创建、修订、发布需要心素权限 feat(prototype)：创建需要心素权限 feat(prototype)：创建角色天赋需要心素权限 feat(prototype)：只有心素才能发起提案、投票提案、删除提案 feat(prototype)：合约api增加赋域/剥夺账号心素令牌，只能由DAO操作 feat(prototype)：合约黑名单在创世区块建立，由DAO管理 revert commit 20116a3：创世区块增加心素nfa定义，同时给创世账号赋予了第一个心素身份 feat：允许向DAO转化YANG为QI test：增加提案相关测试用例，更新各个测试用例

+++ libraries/chain/tps_helper.hpp
@@ -68,9 +68,8 @@ namespace taiyi { namespace chain {
         static ProposalObjectIterator remove_proposal(ProposalObjectIterator& proposal, ProposalIndex& proposalIndex, VotesIndex& votesIndex, const ByVoterIdx& byVoterIdx, tps_removing_reducer& obj_perf)
         {
             // Now remove all votes specific to given proposal.
-            auto propI = byVoterIdx.lower_bound(boost::make_tuple(proposal->id, account_id_type()));
-            
-            while(propI != byVoterIdx.end() && static_cast<size_t>(propI->proposal_id) == static_cast<size_t>(proposal->id))
+            auto propI = byVoterIdx.lower_bound( boost::make_tuple(proposal->id, account_id_type()) );            
+            while(propI != byVoterIdx.end() && propI->proposal_id == proposal->id)
             {
                 auto result_itr = checker< ByProposalType, true/*Loop*/ >(proposal, proposalIndex, obj_perf);
                 if(obj_perf.done)
